+++ CHANGELOG.md
@@ -2,6 +2,7 @@ WIP
 ===
 
 * Support for serving files other than Markdown files. [#6](https://github.com/AustinWise/smeagol/issues/6)
+* Support for specifying which address and port to bind the web server to.
 
 0.1.1 (2022-01-12)
 =====